+++ src/main/java/com/teamtea/eclipticseasons/common/game/SeasonFishingHooks.java
@@ -4,6 +4,7 @@ import com.teamtea.eclipticseasons.api.constant.solar.Season;
 import com.teamtea.eclipticseasons.common.core.biome.WeatherManager;
 import com.teamtea.eclipticseasons.config.CommonConfig;
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
+import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.util.Mth;
@@ -12,14 +13,19 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.LootParams;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParams;
 
+import java.util.List;
+
 public class SeasonFishingHooks {
     public static ObjectArrayList<ItemStack> modify(LootParams pParams, ObjectArrayList<ItemStack> original) {
         if (CommonConfig.Animal.enableFishing.get()) {
             ServerLevel level = pParams.getLevel();
             Entity entity = pParams.getParameter(LootContextParams.THIS_ENTITY);
-            boolean badWeather = WeatherManager.isThunderAt(level, entity.getOnPos().above());
-            Season season = AnimalHooks.getUseSeason(level,entity);
-            if (season != Season.SUMMER || badWeather) {
+            BlockPos blockPos = entity.getOnPos().above();
+            boolean badWeather = WeatherManager.isThunderAt(level, blockPos);
+            Season season = AnimalHooks.getUseSeason(level, entity);
+            if ((season != Season.SUMMER
+                    && (!CommonConfig.Animal.enableCoreWork.get() || AnimalHooks.withoutSeasonBonus(level, blockPos, List.of(Season.SPRING))))
+                    || badWeather) {
                 for (int i = 0; i < original.size(); i++) {
                     var items = original.get(i);
                     if (items.is(ItemTags.FISHES)) {